trying to add map markers
