+++ server/index.html
@@ -7,9 +7,10 @@
     <title>City Guide Backend</title>
 </head>
 <body>
-    <h1>City Guide Backend</h1>
-    <form method="get" action="/api/weather">
-        <input type="submit">
-      </form>
+    <p>STUFF</p>   
+    <script src="./boogey.js"></script>
+    <script>
+        hello();
+    </script>
 </body>
 </html>
