+++ server/location-data.js
@@ -1,0 +1,14 @@
+let myLoc;
+
+function storeLocation(loc) {
+    myLoc = loc;
+    console.log("NEW LOCATION: ", loc);
+}
+
+function getLocation(){
+    return myLoc;
+}
+
+function hello() {
+    console.log("hello");
+}
