+++ server/package.json
@@ -1,8 +1,8 @@
 {
   "dependencies": {
+    "axios": "^1.2.1",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "node-localstorage": "^2.2.1",
     "nodemon": "^2.0.20"
   }
 }
